worked on the search functionality, pagination, started the edit profile page,...

+++ views/partials/table.php
@@ -1,12 +1,12 @@
 <div class="card-group justify-content-center">
-  <table class="table table-stripped table-hover">
+  <table class="table table-hover">
     <tr>
       <th></th>
       <th>Class Name</th>
       <th>Created By</th>
       <th>Date</th>
       <th>
-        <a href="/classes/create" class="btn btn-sm btn-primary "><i class="fa fa-plus"></i>Add New Class</a>
+        Actions
       </th>
     </tr>
 
@@ -25,12 +25,15 @@
             <?= $class['user']['last_name'] ?> <?= $class['user']['first_name'] ?>
           </td>
           <td><?= formatDate($class['createdAt']) ?></td>
-          <td>
-            <a href="/classes/edit?id=<?= $class['id'] ?>" class="btn btn-sm btn-info text-white">
-              <i class="fa fa-edit"></i>
-            </a>
-            <a href="/classes/delete?id=<?= $class['id'] ?>" type="submit" class="btn btn-sm btn-danger"><i class="fa fa-trash-alt"></i></a>
-          </td>
+
+          <?php if (access('lecturer')) : ?>
+            <td>
+              <a href="/classes/edit?id=<?= $class['id'] ?>" class="btn btn-sm btn-info text-white">
+                <i class="fa fa-edit"></i>
+              </a>
+              <a href="/classes/delete?id=<?= $class['id'] ?>" type="submit" class="btn btn-sm btn-danger"><i class="fa fa-trash-alt"></i></a>
+            </td>
+          <?php endif; ?>
         </tr>
       <?php endforeach; ?>
